src/services/backend/requirements.txt
boto3
python-decouple
pandas_datareader
requests
ta
yfinance
pytest
